+++ templates/demo-store/app/data/index.ts
@@ -26,7 +26,7 @@ import type {
 import {type EnhancedMenu, parseMenu, assertApiErrors} from '~/lib/utils';
 import invariant from 'tiny-invariant';
 import {logout} from '~/routes/account/__private/logout';
-import type {AppLoadContext, HydrogenContext} from '@shopify/hydrogen-remix';
+import type {AppLoadContext} from '@remix-run/oxygen';
 
 export interface LayoutData {
   headerMenu: EnhancedMenu;
@@ -35,7 +35,7 @@ export interface LayoutData {
   cart?: Promise<Cart>;
 }
 
-export async function getLayoutData({storefront}: HydrogenContext) {
+export async function getLayoutData({storefront}: AppLoadContext) {
   const HEADER_MENU_HANDLE = 'main-menu';
   const FOOTER_MENU_HANDLE = 'footer';
 
@@ -322,7 +322,7 @@ fragment ImageFragment on Image {
 `;
 
 // shop primary domain url for /admin
-export async function getPrimaryShopDomain({storefront}: HydrogenContext) {
+export async function getPrimaryShopDomain({storefront}: AppLoadContext) {
   const data = await storefront.query<{shop: Shop}>(SHOP_PRIMARY_DOMAIN_QUERY);
 
   invariant(data?.shop?.primaryDomain, 'Primary domain not found');
@@ -734,7 +734,7 @@ const LOGIN_MUTATION = `#graphql
 `;
 
 export async function login(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     email,
     password,
@@ -782,7 +782,7 @@ const CUSTOMER_CREATE_MUTATION = `#graphql
 `;
 
 export async function registerCustomer(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     email,
     password,
@@ -825,7 +825,7 @@ const CUSTOMER_RECOVER_MUTATION = `#graphql
 `;
 
 export async function sendPasswordResetEmail(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {email}: {email: string},
 ) {
   await storefront.mutate<{
@@ -855,7 +855,7 @@ const CUSTOMER_RESET_MUTATION = `#graphql
 `;
 
 export async function resetPassword(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     id,
     resetToken,
@@ -905,7 +905,7 @@ const CUSTOMER_ACTIVATE_MUTATION = `#graphql
 `;
 
 export async function activateAccount(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     id,
     password,
@@ -1131,7 +1131,7 @@ const CUSTOMER_ORDER_QUERY = `#graphql
 `;
 
 export async function getCustomerOrder(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     orderId,
   }: {
@@ -1150,7 +1150,7 @@ export async function getCustomerOrder(
 }
 
 export async function getCustomer(
-  context: AppLoadContext & HydrogenContext,
+  context: AppLoadContext,
   {
     request,
     customerAccessToken,
@@ -1192,7 +1192,7 @@ const CUSTOMER_UPDATE_MUTATION = `#graphql
   `;
 
 export async function updateCustomer(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     customerAccessToken,
     customer,
@@ -1234,7 +1234,7 @@ const UPDATE_ADDRESS_MUTATION = `#graphql
 `;
 
 export async function updateCustomerAddress(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     customerAccessToken,
     addressId,
@@ -1272,7 +1272,7 @@ const DELETE_ADDRESS_MUTATION = `#graphql
 `;
 
 export async function deleteCustomerAddress(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     customerAccessToken,
     addressId,
@@ -1312,7 +1312,7 @@ const UPDATE_DEFAULT_ADDRESS_MUTATION = `#graphql
 `;
 
 export async function updateCustomerDefaultAddress(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     customerAccessToken,
     addressId,
@@ -1355,7 +1355,7 @@ const CREATE_ADDRESS_MUTATION = `#graphql
 `;
 
 export async function createCustomerAddress(
-  {storefront}: HydrogenContext,
+  {storefront}: AppLoadContext,
   {
     customerAccessToken,
     address,