+++ templates/demo-store/app/routes/account.tsx
@@ -24,7 +24,7 @@ import {
   ProductSwimlane,
 } from '~/components';
 import {FeaturedCollections} from '~/components/FeaturedCollections';
-import {type LoaderArgs, redirect, json, defer} from '@shopify/hydrogen-remix';
+import {redirect, json, defer, type LoaderArgs} from '@remix-run/oxygen';
 import {flattenConnection} from '@shopify/hydrogen-react';
 import {getCustomer} from '~/data';
 import {getFeaturedData} from './featured-products';