+++ templates/demo-store/app/routes/account/__public/activate.$id.$activationToken.tsx
@@ -3,8 +3,8 @@ import {
   redirect,
   json,
   type ActionFunction,
-  isStorefrontApiError,
-} from '@shopify/hydrogen-remix';
+} from '@remix-run/oxygen';
+import {isStorefrontApiError} from '@shopify/hydrogen-remix';
 import {Form, useActionData} from '@remix-run/react';
 import {useRef, useState} from 'react';
 import {activateAccount} from '~/data';